+++ src/runtime/io.h
@@ -1,8 +1,11 @@
 #ifndef IO_H
 #define IO_H
 
+#include <stdio.h>
+#include <stdlib.h>
+
 unsigned int read_in_bf();
 
-int read_ws_command_char(FILE *, int *);
+int read_ws_command_char(FILE *, size_t *);
 
 #endif
